Added function static SNode SNode::beautify(SNode) Added function SNode SNode::reduceStmtList(), which does the same as old function SNode::reduce(), SNode::reduce in turn modified to take any SNode with any type/symbol, and call the right reduction function Removed case BOOLCONST from function SNode:reduceRealExp, (irrelevant) Modified drawing of IF-expression, to make a "{", like a fork-expression

+++ Lib/Include/MFCUtil/ExpressionGraphics.h
@@ -56,7 +56,7 @@ public:
   inline int getChildCount() const {
     return (int)m_children.size();
   }
-  inline ExpressionRectangle &child(UINT index) {
+  inline ExpressionRectangle &child(size_t index) {
     return m_children[index];
   }
   inline bool hasParent() const {
